+++ sfml-util/music-set.cpp
@@ -90,7 +90,7 @@ namespace sfml_util
                 }
                 else
                 {
-                    return possibleVec[static_cast<std::size_t>(utilz::random::Int(0, NUM_POSSIBLE_SONGS - static_cast<std::size_t>(1)))];
+                    return possibleVec[static_cast<std::size_t>(utilz::random::Int(0, static_cast<int>(NUM_POSSIBLE_SONGS) - 1))];
                 }
             }
             else
